requirement.txt
google-generativeai
groq
python-dotenv
gitpython
pytest
